Nicer image replacement for #nav-link mobile menu button
The large negative text-indent can sometimes cause horizontal scrollbars
in Right-to-Left languages, and, in some browsers, can also result in a
huge :focus outline being drawn around the button and off to the right
of the window.

Setting a zero height, padding top, and overflow:hidden is a simpler
way to achieve the same thing, and works regardless of text direction,
text length, or focus state.