[Docker] Add required reverse proxy config.
If nginx and fms were on the same machine, ReverseProxy would
automatically be in use, but via docker containers they are not. Do we
need to force it to be switched on? Let's see. There are four possible
options, with their outcome:

* port not in Host, ReverseProxy not in use

Anything using the automatically-generated base instead of BASE_URL uses
port 9000, meaning those links don't work.

* port not in Host, ReverseProxy in use

Anything using the automatically-generated base instead of BASE_URL uses
port 80, meaning those links don't work (they would if you had
docker-compose listen on port 80, being then a similar situation to e.g.
the AMI image).

* port in Host, ReverseProxy not in use

This works *unless* the port is 80, just to be contrary to the above; in
that case it is stripped and :9000 is put back on, meaning those links
again don't work. I realise we use 8000, but would be confusing if
someone tried it out.

* port in Host, ReverseProxy in use

This works in all scenarios, and thus is what we go with.